feat: implement working version of redirect
