+++ src/app/(auth)/chooseRole/RoleButton.tsx
@@ -1,7 +1,5 @@
 'use client';
-
 import Image, { StaticImageData } from 'next/image';
-import photo1 from '@/../public/owner-photo.jpg';
 
 const RoleButton = function ({
   title,
@@ -14,22 +12,19 @@ const RoleButton = function ({
     event.currentTarget.form?.requestSubmit();
   };
 
-  const imageStyle = {
-    borderRadius: '10px'
-  };
-
   return (
-    <button onClick={handleClick} className="w-full shadow-md relative">
-      <Image
-        objectFit="contain"
-        src={photo}
-        alt="Person with her pets"
-        style={imageStyle}
-      />
-      <h2 className="text-brand-fg text-2xl font-bold absolute left-0 right-0 bottom-[10%] m-auto">
-        {title}
-      </h2>
-    </button>
+    <div className="w-full relative bg-gradient-to-b from-brand-fg-200 rounded-xl p-0">
+      <button onClick={handleClick}>
+        <Image
+          src={photo}
+          alt="Person with her pets"
+          className="w-full h-full mix-blend-multiply rounded-xl"
+        />
+        <h2 className="text-brand-fg text-2xl font-bold absolute left-0 right-0 bottom-[10%] m-auto">
+          {title}
+        </h2>
+      </button>
+    </div>
   );
 };
